Instructor and Student signup messages resolved.
Now, only the relevant errors should be visible when an invalid
instructor or student is created.

Also, in the sign up page, prior to this commit there was only
a link for new user that redirected to the instructor's sign up page
but now the new user can sign up as EITHER an new instructor or student!