FINALLY GOT THE PROFILE TAB RENDERING TO WORK!
Now the profile page should give an instuctor the option to launch
his courses from the profile page AS WELL AS add a course
from the profile page.  I added parallel files _new.html.erb
and _show.html.erb to the courses/views because those are
needed for partial rendering (which is what i had to do to
get this work)

Next to do: the Picture setting!