Modifying the instructors profile.
Now the instructor's profile doesn't have an ugly pic
of soloway in the left hand column.  It has instead
two buttons, one for adding a class and one for removing.

The interesting thing about this is however, that
i added to the courses controller a show_all method
for that view.  Then i route the button in my profile
to the courses/show_all page that you guys made.

The interesting thing is that when i click this button
it only shows me the classes that i've made questions
for and activated! No matter who i'm logged in
as.  So if I log in and start and activate a class,
it will be listed as a class I can remove.

However, if I simply "ADD" a class, it will not be
listed as a class I can remove.  It might be because i haven't
modified my db? But i don't think so.

Test it out and lemme know what you think.  i'm not quite sure
what's going on.