Button Icons added to home page.
Now, an icon exists for instructrors to sign up,
students to sign up, and people to learn more on
the home page.

Next, we will add a larger picture to the homepage
where the center hero unit is and perhaps
slightly modify the color scheme.

All pictures are located in public/images
in case you need to locate/modify them!